+++ CHANGELOG.md
@@ -23,6 +23,11 @@
 - [Kirill Kuznetsov] - set correct mode for postgres_exporter executable
 - [Kirill Kuznetsov] - surround attributes for postgresql exporter in quotes.
 
+## 0.1.3
+
+- [Matt Mencel] - CentOS Support: updated only_if and not_if in node_exporter service
+- [Matt Mencel] - Chef Warnings: Fixed Chef warnings in upstart template
+
 ## 0.1.2
 
 - [Kirill Kuznetsov] - PostgreSQL Exporter resource.